+++ sdk/lib/_internal/js_shared/lib/js_interop_patch.dart
@@ -493,14 +493,14 @@ extension JSArrayOfJSNumberToList on JSArray<JSNumber> {
   @patch
   @pragma('dart2js:prefer-inline')
   List<double> get toDartDoubleList => switch (this) {
-    List<double> _ => this,
+    List<double> list => list,
     _ => (this as List).cast<double>(),
   };
 
   @patch
   @pragma('dart2js:prefer-inline')
   List<int> get toDartIntList => switch (this) {
-    List<int> _ => this,
+    List<int> list => list,
     _ => (this as List).cast<int>(),
   };
 }
@@ -519,14 +519,14 @@ extension JSArrayOfNullableJSNumberToList on JSArray<JSNumber?> {
   @patch
   @pragma('dart2js:prefer-inline')
   List<double?> get toDartDoubleList => switch (this) {
-    List<double?> _ => this,
+    List<double?> list => list,
     _ => (this as List).cast<double?>(),
   };
 
   @patch
   @pragma('dart2js:prefer-inline')
   List<int?> get toDartIntList => switch (this) {
-    List<int?> _ => this,
+    List<int?> list => list,
     _ => (this as List).cast<int?>(),
   };
 }
@@ -545,7 +545,7 @@ extension JSArrayOfJSStringToList on JSArray<JSString> {
   @patch
   @pragma('dart2js:prefer-inline')
   List<String> get toDartStringList => switch (this) {
-    List<String> _ => this,
+    List<String> list => list,
     _ => (this as List).cast<String>(),
   };
 }
@@ -563,7 +563,7 @@ extension JSArrayOfNullableJSStringToList on JSArray<JSString?> {
   @patch
   @pragma('dart2js:prefer-inline')
   List<String?> get toDartStringList => switch (this) {
-    List<String?> _ => this,
+    List<String?> list => list,
     _ => (this as List).cast<String?>(),
   };
 }
@@ -580,7 +580,7 @@ extension ListOfNullableStringToJSArray on List<String?> {
 extension JSArrayOfJSBooleanToList on JSArray<JSBoolean> {
   @patch
   List<bool> get toDartBoolList => switch (this) {
-    List<bool> _ => this,
+    List<bool> list => list,
     _ => (this as List).cast<bool>(),
   };
 }
@@ -597,7 +597,7 @@ extension ListOfBoolToJSArray on List<bool> {
 extension JSArrayOfNullableJSBooleanToList on JSArray<JSBoolean?> {
   @patch
   List<bool?> get toDartBoolList => switch (this) {
-    List<bool?> _ => this,
+    List<bool?> list => list,
     _ => (this as List).cast<bool?>(),
   };
 }